Change relative to absolute path
Displaying the absolute path to the audio and video save matches conformity of other paths stated (certs) and clarifies them for docker users who need to map volumes.

NOTE: I don't have subscriptions setup but if I remember correctly there is an /app/subscriptions directory and probably should be mentioned as well. I'll add to this PR if you verify.