Add datestamped tag to nightly docker images
This will make updates to the "nightly" images visible. The `nightly` tag remains as an equivalent of `latest` for development versions, with added `nightly-YYYY-MM-DD` tags for each dev build.

This is generally good practice as using `latest` (or something equivalent) means that two deployments of the same configuration might actually be using different versions of the software. Having immutable tags means it is possible to run a particular development build at will (which might also be useful for testing bugs that only affect some versions, say).

Additionally, this is useful for my particular use case - via the [https://github.com/k8s-at-home/charts](k8s-at-home) project), I've got youtubedl-material deployed in a Kubernetes cluster alongside [https://fluxcd.io/](Flux), which monitors Docker repositories for new image versions with regex pattern matching and automatically updates running deployments to new versions (so when a new nightly build is pushed, my deployment will automatically be updated with it - except I'll be able to tell this has happened because the tag on the image has changed!).